# HumanBenchmarkServer/target/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - HumanBenchmark","contact":{"name":"Raphael .Paquin","url":"https://www.linkedin.com/in/raphee/","email":"raphaelpaquin19@gmail.com"},"version":"1.0"},"servers":[{"url":"http://127.0.0.1:8080"}],"paths":{"/devices":{"get":{"tags":["device-entity-controller"],"description":"get-device","operationId":"getCollectionResource-device-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelDevice"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelDevice"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["device-entity-controller"],"description":"create-device","operationId":"postCollectionResource-device-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelDevice"}}}}}}},"/devices/search/existsByUserName":{"get":{"tags":["device-search-controller"],"operationId":"executeSearch-device-get","parameters":[{"name":"username","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"boolean"}}}},"404":{"description":"Not Found"}}}},"/devices/search/getDeviceByUserName":{"get":{"tags":["device-search-controller"],"operationId":"executeSearch-device-get_1","parameters":[{"name":"username","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelDevice"}}}},"404":{"description":"Not Found"}}}},"/devices/{id}":{"get":{"tags":["device-entity-controller"],"description":"get-device","operationId":"getItemResource-device-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelDevice"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["device-entity-controller"],"description":"update-device","operationId":"putItemResource-device-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelDevice"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelDevice"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["device-entity-controller"],"description":"delete-device","operationId":"deleteItemResource-device-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["device-entity-controller"],"description":"patch-device","operationId":"patchItemResource-device-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelDevice"}}}},"204":{"description":"No Content"}}}},"/games":{"get":{"tags":["game-entity-controller"],"description":"get-game","operationId":"getCollectionResource-game-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelGame"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelGame"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["game-entity-controller"],"description":"create-game","operationId":"postCollectionResource-game-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelGame"}}}}}}},"/games/{id}":{"get":{"tags":["game-entity-controller"],"description":"get-game","operationId":"getItemResource-game-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelGame"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["game-entity-controller"],"description":"update-game","operationId":"putItemResource-game-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelGame"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelGame"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["game-entity-controller"],"description":"delete-game","operationId":"deleteItemResource-game-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["game-entity-controller"],"description":"patch-game","operationId":"patchItemResource-game-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelGame"}}}},"204":{"description":"No Content"}}}},"/lobbies":{"get":{"tags":["lobby-entity-controller"],"description":"get-lobby","operationId":"getCollectionResource-lobby-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelLobby"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelLobby"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["lobby-entity-controller"],"description":"create-lobby","operationId":"postCollectionResource-lobby-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LobbyRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelLobby"}}}}}}},"/lobbies/search/existsBySessionCode":{"get":{"tags":["lobby-search-controller"],"operationId":"executeSearch-lobby-get","parameters":[{"name":"sessionCode","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"boolean"}}}},"404":{"description":"Not Found"}}}},"/lobbies/search/getSessionBySessionCode":{"get":{"tags":["lobby-search-controller"],"operationId":"executeSearch-lobby-get_1","parameters":[{"name":"sessionCode","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelLobby"}}}},"404":{"description":"Not Found"}}}},"/lobbies/{id}":{"get":{"tags":["lobby-entity-controller"],"description":"get-lobby","operationId":"getItemResource-lobby-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelLobby"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["lobby-entity-controller"],"description":"update-lobby","operationId":"putItemResource-lobby-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LobbyRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelLobby"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelLobby"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["lobby-entity-controller"],"description":"delete-lobby","operationId":"deleteItemResource-lobby-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["lobby-entity-controller"],"description":"patch-lobby","operationId":"patchItemResource-lobby-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LobbyRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelLobby"}}}},"204":{"description":"No Content"}}}},"/lobbies/{id}/players":{"get":{"tags":["lobby-property-reference-controller"],"description":"get-device-by-lobby-Id","operationId":"followPropertyReference-lobby-get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelDevice"}},"text/uri-list":{"schema":{"type":"string"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["lobby-property-reference-controller"],"description":"update-device-by-lobby-Id","operationId":"createPropertyReference-lobby-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelDevice"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelDevice"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["lobby-property-reference-controller"],"description":"delete-device-by-lobby-Id","operationId":"deletePropertyReference-lobby-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["lobby-property-reference-controller"],"description":"patch-device-by-lobby-Id","operationId":"createPropertyReference-lobby-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelDevice"}}}},"204":{"description":"No Content"}}}},"/lobbies/{id}/players/{propertyId}":{"get":{"tags":["lobby-property-reference-controller"],"description":"get-device-by-lobby-Id","operationId":"followPropertyReference-lobby-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelDevice"}}}},"404":{"description":"Not Found"}}},"delete":{"tags":["lobby-property-reference-controller"],"description":"delete-device-by-lobby-Id","operationId":"deletePropertyReferenceId-lobby-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}}},"/profile":{"get":{"tags":["profile-controller"],"operationId":"listAllFormsOfMetadata_1","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/RepresentationModelObject"}}}}}}},"/profile/devices":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/profile/games":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/profile/lobbies":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/v1/device/register-device":{"post":{"tags":["device-controller"],"summary":"Registers a device.","operationId":"registerDevice","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"422":{"description":"Device Already Exists."},"400":{"description":"Bad Request."},"200":{"description":"Registers a device.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}}}}},"/v1/lobby/get-lobby":{"get":{"tags":["lobby-controller"],"summary":"Gets a lobby's players.","operationId":"getLobby","parameters":[{"name":"lobbyCode","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request."},"404":{"description":"Lobby not found."},"200":{"description":"Fetched lobby information.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Device"}}}}}}}},"/v1/lobby/create-lobby":{"get":{"tags":["lobby-controller"],"summary":"Creates a lobby.","operationId":"createLobby","parameters":[{"name":"deviceName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request."},"200":{"description":"Creates a lobby.","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"AbstractJsonSchemaPropertyObject":{"type":"object","properties":{"title":{"type":"string"},"readOnly":{"type":"boolean"}}},"Item":{"type":"object","properties":{"type":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}}}},"JsonSchema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}},"definitions":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Item"}},"type":{"type":"string"},"$schema":{"type":"string"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"RepresentationModelObject":{"type":"object","properties":{"_links":{"$ref":"#/components/schemas/Links"}}},"EntityModelGame":{"type":"object","properties":{"gameName":{"type":"string"},"gameDescription":{"type":"string"},"gameType":{"type":"string","enum":["CHIMP_GAME","VISUAL_MEMORY_GAME","VERBAL_MEMORY_GAME","DIGIT_MEMORY_GAME"]},"_links":{"$ref":"#/components/schemas/Links"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelGame":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"games":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelGame"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"Device":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userName":{"type":"string"},"donePlaying":{"type":"boolean"},"score":{"type":"integer","format":"int32"}}},"EntityModelLobby":{"type":"object","properties":{"sessionCode":{"type":"string"},"gameCompleted":{"type":"boolean"},"isStarted":{"type":"boolean"},"_links":{"$ref":"#/components/schemas/Links"}}},"PagedModelEntityModelLobby":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"lobbies":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelLobby"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"CollectionModelDevice":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"devices":{"type":"array","items":{"$ref":"#/components/schemas/DeviceResponse"}}}},"_links":{"$ref":"#/components/schemas/Links"}}},"CollectionModelObject":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"objects":{"type":"array","items":{"type":"object"}}}},"_links":{"$ref":"#/components/schemas/Links"}}},"EntityModelDevice":{"type":"object","properties":{"userName":{"type":"string"},"donePlaying":{"type":"boolean"},"score":{"type":"integer","format":"int32"},"_links":{"$ref":"#/components/schemas/Links"}}},"PagedModelEntityModelDevice":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"devices":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelDevice"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"DeviceRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userName":{"type":"string"},"donePlaying":{"type":"boolean"},"score":{"type":"integer","format":"int32"}}},"GameRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"gameName":{"type":"string"},"gameDescription":{"type":"string"},"gameType":{"type":"string","enum":["CHIMP_GAME","VISUAL_MEMORY_GAME","VERBAL_MEMORY_GAME","DIGIT_MEMORY_GAME"]}}},"LobbyRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sessionCode":{"type":"string"},"players":{"type":"array","items":{"type":"string"}},"gameCompleted":{"type":"boolean"},"isStarted":{"type":"boolean"}}},"DeviceResponse":{"type":"object","properties":{"userName":{"type":"string"},"donePlaying":{"type":"boolean"},"score":{"type":"integer","format":"int32"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}